implement audition-scheduling
The complete scheduling of auditions is kind of done. It requires that the
database has registered users with selected big and small pleasures.
It then needs data.csv which is the output from the google form. And it parses
the csv file.

It also has a configuration file called group.json. Where configurations
on how the different 'juries' work.

The code outputs two csv files. One for the jury so that they can see their
schedule. And one for the user. Which can be used to mass-mail the users with
a gs-script on google spreadsheet.

Magic. So much magic.'