Add Dockerfiles for dockerization
For those who want to easily start a database and backend with `docker-compose up`.
Could even in the future contain karnevalist.se, and therefore we could create the
whole environment (frontend and backend) with docker. Cool!

For not having problem with bcrypt binary i changed it to a specific version instead
of version 1.0.3 or more.